requirements.txt
django
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
celery==5.5.3
redis==6.2.0
django-celery-beat==2.8.1
psycopg==3.2.9